+++ src/Components/Navbar/NavBar.js
@@ -1,14 +1,15 @@
 import React, { Component } from 'react';
 import {Route, NavLink} from 'react-router-dom';
-import Landing from './Landing';
-import Card from './Card';
-import mockPeopleData from './mockPeopleData'
-import mockPlanetsData from './mockPlanetsData'
-import mockVehicleData from './mockVehiclesData'
+import Landing from '../Landing/Landing';
+import Card from '../Card/Card';
+import mockPeopleData from '../../Data/mockPeopleData'
+import mockPlanetsData from '../../Data/mockPlanetsData'
+import mockVehicleData from '../../Data/mockVehiclesData'
 import './NavBar.css'
-import r2d2 from './r2d2.svg'
-import mFalcon from './millennium-falcon.svg'
-import dStar from './death-star.svg'
+import r2d2 from '../../Images/r2d2.svg'
+import mFalcon from '../../Images/millennium-falcon.svg'
+import dStar from '../../Images/death-star.svg'
+
 
 
 class NavBar extends Component {
@@ -31,7 +32,7 @@ class NavBar extends Component {
           <Route exact path='/' component={Landing} />
         <section className='card-container'>
           <Route exact path='/People' render={() => <Card data={mockPeopleData.results}/>} />
-          <Route exact path='/Planets' render={() => <Card data={mockPlanetsData.results}/>} />          
+          <Route exact path='/Planets' render={() => <Card data={mockPlanetsData.results}/>} />
           <Route exact path='/Vehicles' render={() => <Card data={mockVehicleData.results}/>} />
         </section>
       </main>
