+++ src/Components/ScrollText/ScrollText.js
